requirements.txt
fastapi
uvicorn[standard]
redis
rq
psycopg2-binary
sqlalchemy
sqlalchemy[asyncio]
asyncpg
pydantic
python-dotenv